Group resource into data package types
