+++ frictionless/resources/__init__.py
@@ -1,9 +1,5 @@
 from .file import *
-from .inquiry import *
 from .json import *
-from .metadata import *
-from .package import *
-from .resource import *
 from .table import *
 from .text import *
 from .types import *
